[Modify]技数をバッチスタイルで右上に表示
